// apps/patient-web/src/types/models.ts
export type ApiError = {
    message: string
    status?: number
    errors?: unknown
}